Outras exceções

+++ src/main/java/com/novaes/NovaesCommerce/services/ProductService.java
@@ -4,38 +4,38 @@
  */
 package com.novaes.NovaesCommerce.services;
 
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.Pageable;
+import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Propagation;
+import org.springframework.transaction.annotation.Transactional;
+
 import com.novaes.NovaesCommerce.dto.ProductDTO;
 import com.novaes.NovaesCommerce.entities.Product;
 import com.novaes.NovaesCommerce.repositories.ProductRepository;
-import com.novaes.NovaesCommerce.services.exceptions.ResorceNotFoundException;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.data.domain.Page;
-import org.springframework.data.domain.Pageable;
+import com.novaes.NovaesCommerce.services.exceptions.DatabaseException;
+import com.novaes.NovaesCommerce.services.exceptions.ResourceNotFoundException;
 
-import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Transactional;
+import jakarta.persistence.EntityNotFoundException;
 
 /**
  *
  * @author blnov
  */
+
 @Service
 public class ProductService {
 
     @Autowired
     private ProductRepository repository;
 
-    /**
-     *
-     * @param id
-     * @return
-     */
     @Transactional(readOnly = true)
     public ProductDTO findById(Long id) {
-            Product product = repository.findById(id)
-                    .orElseThrow(()-> new ResorceNotFoundException("Recurso não encontrado"));
-            return new ProductDTO(product);
-
+        Product product = repository.findById(id).orElseThrow(
+                () -> new ResourceNotFoundException("Recurso não encontrado"));
+        return new ProductDTO(product);
     }
 
     @Transactional(readOnly = true)
@@ -44,34 +44,44 @@ public class ProductService {
         return result.map(x -> new ProductDTO(x));
     }
 
-    @Transactional()
+    @Transactional
     public ProductDTO insert(ProductDTO dto) {
         Product entity = new Product();
         copyDtoToEntity(dto, entity);
         entity = repository.save(entity);
-
         return new ProductDTO(entity);
     }
 
     @Transactional
     public ProductDTO update(Long id, ProductDTO dto) {
-        Product entity = repository.getReferenceById(id);
-        copyDtoToEntity(dto, entity);
-        entity = repository.save(entity);
-
-        return new ProductDTO(entity);
+        try {
+            Product entity = repository.getReferenceById(id);
+            copyDtoToEntity(dto, entity);
+            entity = repository.save(entity);
+            return new ProductDTO(entity);
+        }
+        catch (EntityNotFoundException e) {
+            throw new ResourceNotFoundException("Recurso não encontrado");
+        }
     }
 
-    @Transactional
+    @Transactional(propagation = Propagation.SUPPORTS)
     public void delete(Long id) {
-        repository.deleteById(id);
+    	if (!repository.existsById(id)) {
+    		throw new ResourceNotFoundException("Recurso não encontrado");
+    	}
+    	try {
+            repository.deleteById(id);    		
+    	}
+        catch (DataIntegrityViolationException e) {
+            throw new DatabaseException("Falha de integridade referencial");
+        }
     }
 
     private void copyDtoToEntity(ProductDTO dto, Product entity) {
         entity.setName(dto.getName());
         entity.setDescription(dto.getDescription());
-        entity.setImgUri(dto.getImgUrl());
         entity.setPrice(dto.getPrice());
+        entity.setImgUrl(dto.getImgUrl());
     }
-
 }
